refactor : Progress 패키지 이동
